requirements.txt
asgiref==3.6.0
black==22.1.0
blue==0.9.1
click==8.1.3
colorama==0.4.6
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==4.0.1
mccabe==0.6.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
psycopg2==2.9.6
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3
